Refactor the data handling modul
