+++ answer.csv
@@ -5,3 +5,7 @@ id,submission_time,vote_number,question_id,message,image
 3,1571744676,0,0,Nothing TROLOLOLO,
 4,1571746051,0,0,Try to look it up on google,
 5,1571749594,0,0,New answer,
+6,1571818444,,7,Just chekin',
+7,1571820548,,3,Szerintem ja,
+8,1571821033,,7,new answer,
+9,1571821124,0,7,cheking 0 vote value,
