+++ data_handler.py
@@ -1,54 +1,23 @@
 import csv
-import time
+import connection
+
 
 ANSWER_DATA_PATH = 'answer.csv'
 QUESTION_DATA_PATH = 'question.csv'
-ANSWER_HEADERS = ['id', 'submission_time','vote_number', 'question_id', 'message','image']
-QUESTION_HEADERS = ['id', 'submission_time', 'view_number','vote_number', 'title', 'message', 'image']
+ANSWER_HEADERS = ['id', 'submission_time', 'vote_number', 'question_id', 'message', 'image']
+QUESTION_HEADERS = ['id', 'submission_time', 'view_number', 'vote_number', 'title', 'message', 'image']
 
 
-def get_data_from_csv(csv_file):
-    with open(csv_file, encoding='utf-8') as file:
-        reader = csv.DictReader(file)
-        all_data = list(reader)
-        for row in all_data:
-            for key, val in row.items():
-                if is_number(val):
-                    row[key] = int(val)
-        return all_data
-
-
-def is_number(s):
-    try:
-        int(s)
-        return True
-    except ValueError:
-        pass
-    return False
-
-
-def get_all_data(csv_file, break_lines=False):
-    all_data = get_data_from_csv(csv_file)
-    if break_lines:
-        for data in all_data:
-            data['message'] = convert_enter_to_br(data['message'])
-    return all_data
-
-
-def convert_enter_to_br(original_string):
-    return '<br>'.join(original_string.split('\n'))
-
-
-def creat_new_id(csv_file):
-    existing_data = get_data_from_csv(csv_file)
-    if len(existing_data) == 0:
-        return'1'
-    return str(int(existing_data[-1]['id'])+1)
+def get_data(data_type):
+    if data_type == 'questions':
+        return connection.get_all_data(QUESTION_DATA_PATH)
+    elif data_type == 'answers':
+        return connection.get_all_data(ANSWER_DATA_PATH)
 
 
 def add_new_question(question):
-    question['id'] = creat_new_id(QUESTION_DATA_PATH)
-    question['submission_time'] = get_submission_time()
+    question['id'] = connection.creat_new_id(QUESTION_DATA_PATH)
+    question['submission_time'] = connection.get_submission_time()
     question['vote_number'] = 0
     question['view_number'] = 0
     add_new_data_to_file(question, QUESTION_DATA_PATH, QUESTION_HEADERS)
@@ -62,14 +31,11 @@ def add_new_data_to_file(data, file_to, header):
 
 def add_new_message(answer, question_id):
     answer = dict(answer)
-    answer['id'] = creat_new_id(ANSWER_DATA_PATH)
-    answer['submission_time'] = get_submission_time()
+    answer['id'] = connection.creat_new_id(ANSWER_DATA_PATH)
+    answer['submission_time'] = connection.get_submission_time()
     answer['question_id'] = question_id
+    answer['vote_number'] = 0
     add_new_data_to_file(answer, ANSWER_DATA_PATH, ANSWER_HEADERS)
-
-
-def get_submission_time():
-    return int(time.time())
 
 
 def edit_question(data, file_to):
@@ -86,3 +52,10 @@ def update_existing_file(data, file_to, header):
         writer.writeheader()
         for row in data:
             writer.writerow(row)
+
+
+def get_row_index_by_id(id_to_find, data):
+    id_to_find = int(id_to_find)
+    for index, row in enumerate(data):
+        if row['id'] == id_to_find:
+            return index
