+++ question.csv
@@ -1,8 +1,9 @@
 id,submission_time,view_number,vote_number,title,message,image
 0,1493368154,29,7,Python lists,"How to define lists in python?<br>I am TOTALLY new to this, please help!!!!!!!",
-1,1493068124,15,9,Wordpress loading multiple jQuery Versions,"I developed a plugin that uses the jquery booklet plugin (http://builtbywill.com/booklet/#/) this plugin binds a function to $ so I cann call $('.myBook').booklet();<br><br>I could easy managing the loading order with wp_enqueue_script so first I load jquery then I load booklet so everything is fine.<br><br>BUT in my theme i also using jquery via webpack so the loading order is now following:<br><br>jquery<br>booklet<br>app.js (bundled file with webpack, including jquery)",images/image1.png
+1,1493068124,15,8,Wordpress loading multiple jQuery Versions,"I developed a plugin that uses the jquery booklet plugin (http://builtbywill.com/booklet/#/) this plugin binds a function to $ so I cann call $('.myBook').booklet();<br><br>I could easy managing the loading order with wp_enqueue_script so first I load jquery then I load booklet so everything is fine.<br><br>BUT in my theme i also using jquery via webpack so the loading order is now following:<br><br>jquery<br>booklet<br>app.js (bundled file with webpack, including jquery)",images/image1.png
 2,1493015432,1364,57,Drawing canvas with an image picked with Cordova Camera Plugin,"I'm getting an image from device and drawing a canvas with filters using Pixi JS. It works all well using computer to get an image. But when I'm on IOS, it throws errors such as cross origin issue, or that I'm trying to use an unknown format.<br><br>This is the code I'm using to draw the image (that works on web/desktop but not cordova built ios app)",
-3,1571658393,0,9,Működik ez?,Működik?<br>NA?<br>NAAAAAAAAAAAAAAAAAAAA?,
-4,1571659447,0,4,asdasd,awesxrtfvgbhunjim,
+3,1571658393,0,8,Működik ez???,Működik?,
+4,1571659447,0,2,asdasd,awesxrtfvgbhunjim,
 5,1571660377,0,-2,werser,sadasd,
-6,1571747851,0,1,Is the vote function works?,Down vote up vote dose not matter,
+6,1571747851,0,2,Is the vote function works?,Down vote up vote dose not matter,
+7,1571818414,0,2,Just cheking v.3.0,Just cheking,
