+++ server.py
@@ -7,17 +7,10 @@ app = Flask(__name__)
 saved_data = {}
 
 
-def get_row_index_by_id(id_to_find, data):
-    id_to_find = int(id_to_find)
-    for index, row in enumerate(data):
-        if row['id'] == id_to_find:
-            return index
-
-
 @app.route('/')
 @app.route('/list')
 def route_list():
-    questions = data_handler.get_all_data('question.csv',break_lines=True)
+    questions = data_handler.get_data('questions')
     should_reverse = True
     if request.args.get('order_direction') == 'asc':
         should_reverse = False
@@ -44,9 +37,9 @@ def add_question():
 @app.route('/question/<question_id>')
 def show_answers(question_id):
     question_id = int(question_id)
-    questions = data_handler.get_all_data('question.csv', True)
+    questions = data_handler.get_data('questions')
     title = ''.join([question['title'] for question in questions if question['id'] == question_id])
-    answers = data_handler.get_all_data('answer.csv', True)
+    answers = data_handler.get_data('answers')
     answers_for_question = [answer for answer in answers if answer['question_id'] == question_id]
     return render_template('answer.html',
                            question_title=title,
@@ -59,20 +52,20 @@ def add_new_answer(question_id):
     if request.method == 'POST':
         data_handler.add_new_message(request.form, question_id)
         return redirect(f'/question/{question_id}')
-    questions = data_handler.get_all_data('question.csv', True)
+    questions = data_handler.get_data('questions')
     title = ''.join([question['title'] for question in questions if question['id'] == question_id])
     return render_template('new_answer.html', title=title, question_id=question_id)
 
 
 @app.route('/question/<question_id>/edit', methods=['GET','POST'])
 def edit_question(question_id):
-    all_questions = data_handler.get_data_from_csv("question.csv")
-    question_index = get_row_index_by_id(question_id, all_questions)
+    all_questions = data_handler.get_data('questions')
+    question_index = data_handler.get_row_index_by_id(question_id, all_questions)
     question_data = all_questions[question_index]
     if request.method == 'POST':
         question_data['title'] = request.form.get('title')
         question_data['message'] = request.form.get('message')
-        data_handler.update_existing_file(all_questions, 'question.csv', data_handler.QUESTION_HEADERS)
+        data_handler.update_existing_file(all_questions, data_handler.QUESTION_DATA_PATH, data_handler.QUESTION_HEADERS)
         return redirect('/list')
     return render_template('edit_question.html',
                            title='Edit Question',
@@ -82,7 +75,7 @@ def edit_question(question_id):
 
 @app.route('/question/<question_id>/vote-up')
 def question_vote_up(question_id):
-    questions = data_handler.get_all_data('question.csv', True)
+    questions = data_handler.get_data('questions')
     for question in questions:
         if question['id'] == int(question_id):
             question['vote_number'] = question['vote_number'] + 1
@@ -92,7 +85,7 @@ def question_vote_up(question_id):
 
 @app.route('/question/<question_id>/vote-down')
 def question_vote_down(question_id):
-    questions = data_handler.get_all_data('question.csv', True)
+    questions = data_handler.get_data('questions')
     for question in questions:
         if question['id'] == int(question_id):
             question['vote_number'] = question['vote_number'] - 1
